Swapped out GSL integrator for integration::QAG

+++ src/control.h
@@ -22,7 +22,7 @@ public:
   bool   plant_assimilation_over_distribution;
   double plant_assimilation_tol;
   size_t plant_assimilation_iterations;
-  int    plant_assimilation_rule;
+  size_t plant_assimilation_rule;
 
   double plant_seed_tol;
   int    plant_seed_iterations;
